requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudinary==1.31.0
colorama==0.4.6
dj-database-url==1.2.0
Django==4.1.5
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
